student-dropout-prediction/requirements.txt
streamlit
pandas
numpy
plotly
requests
Pillow
Flask
gunicorn
opencv-python-headless
Flask
Flask-Cors
tensorflow
